+++ src/Crack.cpp
@@ -6,7 +6,7 @@ Crack::Crack() : Ground()
 
 }
 
-Crack::Crack(int x, int y, int z) : Ground(x,y,z)
+Crack::Crack(GLfloat x, GLfloat y, GLfloat z) : Ground(x,y,z)
 {
     this->setModel("Cube.obj");
 }
