+++ src/Floor.cpp
@@ -6,7 +6,7 @@ Floor::Floor() : Ground()
 
 }
 
-Floor::Floor(int x, int y, int z) : Ground(x,y,z)
+Floor::Floor(GLfloat x, GLfloat y, GLfloat z) : Ground(x,y,z)
 {
     this->setModel("Cube.obj");
 }
